Clean up theory:
- move functions used in testing to test file
- move plotting to plotting_tools and anchor generation to measurements
- add some documentation
- add run_notebooks instalation instructions for conda
