+++ scripts/run_notebooks
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-runner.py: Test if notebooks run without error.  
+run_notebooks.py: Test if notebooks run without error.
 """
 
 from nbconvert.preprocessors import ExecutePreprocessor
